+++ lib/bindables/depth_stencil/depth_stencil_view.h
@@ -13,7 +13,7 @@ namespace PULSAR
 		ID3D11DepthStencilView *mp_ds_view = NULL;
 
 	private:
-		void	create_ds_view(float width, float height);
+		void	create_ds_view(UINT width, UINT height);
 		void	free();
 
 	public:
@@ -22,10 +22,10 @@ namespace PULSAR
 		depth_stencil_view(const depth_stencil_view &) = delete;
 		depth_stencil_view(depth_stencil_view&&) = delete;
 		depth_stencil_view() = default;
-		depth_stencil_view(float width, float height);
+		depth_stencil_view(UINT width, UINT height);
 		~depth_stencil_view();
 
-		void	set(float width, float height);
+		void	set(UINT width, UINT height);
 		ID3D11Texture2D	*get_texture();
 		ID3D11DepthStencilView	*get_view();
