+++ lib/bindables/viewport/viewport.h
@@ -11,7 +11,7 @@ namespace PULSAR
 		D3D11_VIEWPORT m_viewport = { 0 };
 
 	private:
-		void	create_viewport(float width, float height, float min_depth, float max_depth);
+		void	create_viewport(UINT width, UINT height, float min_depth, float max_depth);
 
 	public:
 		viewport &operator=(const viewport&) = delete;
@@ -19,10 +19,10 @@ namespace PULSAR
 		viewport(const viewport&) = delete;
 		viewport(viewport&&) = delete;
 		viewport() = default;
-		viewport(float width, float height, float min_depth = 0.0f, float max_depth = 1.0f);
+		viewport(UINT width, UINT height, float min_depth = 0.0f, float max_depth = 1.0f);
 		~viewport() = default;
 
-		void	set(float width, float height, float min_depth = 0.0f, float max_depth = 1.0f);
+		void	set(UINT width, UINT height, float min_depth = 0.0f, float max_depth = 1.0f);
 		D3D11_VIEWPORT	&get();
 
 		void	bind() const override;
