+++ lib/bindables/viewport/viewport_data.cpp
@@ -2,7 +2,7 @@
 #include "viewport.h"
 
 
-void	PULSAR::viewport::create_viewport(float width, float height, float min_depth, float max_depth)
+void	PULSAR::viewport::create_viewport(UINT width, UINT height, float min_depth, float max_depth)
 {
 	this->m_viewport.TopLeftX = 0.0f;
 	this->m_viewport.TopLeftY = 0.0f;
@@ -12,7 +12,7 @@ void	PULSAR::viewport::create_viewport(float width, float height, float min_dept
 	this->m_viewport.MaxDepth = max_depth;
 }
 
-void	PULSAR::viewport::set(float width, float height, float min_depth, float max_depth)
+void	PULSAR::viewport::set(UINT width, UINT height, float min_depth, float max_depth)
 {
 	this->create_viewport(width, height, min_depth, max_depth);
 }
